fixed some typos in comments now using relative paths for embedded assets in sample projects
That makes it easier to compile them in other IDEs (like IntelliJ) because it may avoid having to modify the source paths.
added helper script for IntelliJ debugging

The Ruby script provides a workaround for a bug in IntelliJ: it does not copy the specified resources into the output-directory when running/debugging the app in the simulator. Configured as an external tool, the script can do just that.

-> https://youtrack.jetbrains.com/issue/IDEA-94578